backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
google-genai==1.2.0
numpy==1.26.3
scikit-learn==1.4.0
python-multipart==0.0.6
httpx==0.26.0
